Explicitly annotating [Scope], [AssertedScope], [IdentifierName], [PropertyKey].
Encoders, decoders and tooling need more precise information than the purely syntactic `IdentifierName`.
In particular, the three following categories have very different properties with respect to encoding
strategies, possible renamings, scope information:
- string property names;
- string identifier declarations;
- string identifier references.

So, we replace the sole `IdentifierName` with two distinct typedefs `PropertyKey` and `IdentifierName`
and their companion hints `[PropertyKey]` and`[IdentifierName]`.

Moreover, since there is no trivial definition of the scope for which an identifier declaration is
valid, we add a `[Scope]` annotation that specifies scope boundaries for any `[IdentifierName]`
it contains and a `[AssertedScope]` that specifies where an `[IdentifierName]` is declared.

We expect these annotations to be useful both for encoding (e.g. to determine per-scope
encoding strategies) and for tooling (e.g. to aid debugging tools or refactoring tools trace variable
information more easily).